refactor : useFilter 병합
