+++ src/hooks/index.js
@@ -1,3 +1,3 @@
 export * from "./useInfiniteList";
 export * from "./useIntersect";
-export * from "./useFilterCategory";
+export * from "./useFilter";
